+++ BunnyMunch/src/utopia/bunnymunch/jawn/JsonBunny.scala
@@ -4,7 +4,10 @@ import java.io.{File, InputStream}
 import java.nio.file.Path
 
 import org.typelevel.jawn.Parser
+import utopia.flow.datastructure.immutable.Value
 import utopia.flow.util.StringFrom
+import utopia.flow.generic.ValueConversions._
+import utopia.flow.parse.JsonParser
 
 import scala.io.Codec
 
@@ -14,15 +17,33 @@ import scala.io.Codec
   * @author Mikko Hilpinen
   * @since 12.5.2020, v
   */
-object JsonBunny
+object JsonBunny extends JsonParser
 {
 	// ATTRIBUTES	----------------------------
 	
-	private implicit val defaultEncoding: Codec = Codec.UTF8
+	/**
+	  * The encoding accepted by this parser (currently only supports utf-8)
+	  */
+	implicit val defaultEncoding: Codec = Codec.UTF8
 	private implicit val facade: ValueFacade.type = ValueFacade
 	
 	
+	// IMPLEMENTED	----------------------------
+	
+	override def apply(json: String) = munch(json)
+	
+	override def apply(file: File) = munchFile(file)
+	
+	override def apply(inputStream: InputStream) = munchStream(inputStream)
+	
+	
 	// OTHER	--------------------------------
+	
+	/**
+	  * @param s A string possibly representing a json value
+	  * @return Json value of the string, if it could be parsed. If not, simply returns the value as a string.
+	  */
+	def sureMunch(s: String) = munch(s).toOption.getOrElse(s: Value)
 	
 	/**
 	  * @param json Json to parse