+++ Nexus/Changes.md
@@ -1,4 +1,8 @@
 # Utopia Nexus - List of Changes
-## v1.5
+## v1.5 (Beta)
 ### Breaking Changes
 - Renamed NoOperation result to NotModified
+### Deprecations
+- StreamedBody.bufferedJSON, .bufferedJSONModel and .bufferedJSONArray were deprecated in favor of new 
+.bufferedJson, .bufferedJsonObject and .bufferedJsonArray. The new implementations take an implicit 
+JsonParser, so they are no longer locked to JSONReader only.